Move components into 'components' folder
I realized these should probably be in a components folder, so
let's do that.
