+++ packages/ui/footer/components/Legal.tsx
