+++ packages/ui/footer/components/Newsletter.tsx
