+++ packages/ui/footer/components/Partners.tsx
